feat: Added 2 projects, dll and exe.
The dll is right now not set up correctly and you need the Catch.dll in the directory where your code is located in the other project. I have created a test namespace to check if the dll is working, and it is working. Using __declspec(dllimport) and __declspec(dllexport) I am making sure that the function in the namespace is correctly exported by the Catch.dll and imported by main.cpp in the Sandbox project.