feat: Added Premake to make life easier and compile code using the bat file
Using Premake as it uses Lua and is bit simpler than CMake. Making 3 configuration builds release, debug, Dist for the project. Adding all the defines and include folders we need in the Premake file. Due to premake, removed the need for copying the Catch.dll to the Sandbox bin folder again and agian, it will be done automatically by the postbuildcommand in the premake file. You may need to compile the code twice for its proper functioning. The comments have been added to the premake file for understanding. Learnt to create bat files and write commands in them as well as learnt how to write good premake files for our project.